requirements.txt
diffusers
transformers
accelerate
#torch
#xformers
gradio==3.50.2
Pillow
qrcode
